Handle the binary functions too
